+++ locomotion-sdk/src/pages/AuthScreens/Onboarding/Phone.js
@@ -1,5 +1,4 @@
 import React, { useState } from 'react';
-import PhoneInput from 'react-native-phone-number-input';
 import { useNavigation } from '@react-navigation/native';
 import i18n from '../../../I18n';
 import OnboardingNavButtons from './OnboardingNavButtons';
@@ -8,15 +7,14 @@ import { ErrorText, PageContainer, SafeView } from './styles';
 import Header from './Header';
 import ScreenText from './ScreenText/index';
 import { loginApi } from '../../../context/user/api';
-import { ERROR_COLOR } from '../../../services/sharedStyles';
+import PhoneNumberInput from '../../../Components/PhoneNumberInput';
 
 const Phone = () => {
   const { onboardingState, updateState } = onboardingContext.useContainer();
   const navigation = useNavigation();
   const [showErrorText, setShowErrorText] = useState(false);
-  const [isFocused, setIsFocused] = useState(false);
-  const [countryCode, setCountryCode] = useState('972');
-  const onPhoneNumberChange = (phoneNumber) => {
+
+  const onPhoneNumberChange = (phoneNumber, countryCode) => {
     setShowErrorText(false);
     if (phoneNumber.length < 9) {
       return updateState('phoneNumber', '');
@@ -36,10 +34,6 @@ const Phone = () => {
     }
   };
 
-  const onChangeCountry = (v) => {
-    setCountryCode(v.callingCode[0]);
-  };
-
   return (
     <SafeView>
       <Header title={i18n.t('onboarding.pages.phone.title')} />
@@ -48,33 +42,11 @@ const Phone = () => {
           text={i18n.t('onboarding.pages.phone.text')}
           subText={i18n.t('onboarding.pages.phone.subText')}
         />
-        <PhoneInput
+        <PhoneNumberInput 
+          onPhoneNumberChange={onPhoneNumberChange}
           autoFocus
           defaultCode="IL"
-          onChangeText={onPhoneNumberChange}
-          textInputProps={{
-            onFocus: () => setIsFocused(true),
-            onBlur: () => setIsFocused(false),
-          }}
-          onChangeCountry={onChangeCountry}
-          containerStyle={{
-            width: '100%',
-          }}
-          placeholder={i18n.t('onboarding.pages.phone.placeholder')}
-          textContainerStyle={{
-            borderRadius: 8,
-            backgroundColor: '#f1f2f6',
-            borderWidth: isFocused ? 0.5 : 0,
-            borderColor: showErrorText ? ERROR_COLOR : '#333333',
-          }}
-          textInputStyle={{
-            color: showErrorText ? ERROR_COLOR : '#333333',
-          }}
-          flagButtonStyle={{
-            backgroundColor: '#f1f2f6',
-            borderRadius: 8,
-            marginRight: 4,
-          }}
+          error={showErrorText}
         />
         {showErrorText && <ErrorText>{showErrorText}</ErrorText>}
         <OnboardingNavButtons
